refactor: Move functions to get WikiData to separate file
