+++ src/luchtmans/views.py
@@ -10,20 +10,7 @@ from django.utils import translation
 
 from .models import Country, Person, Place
 from .utils import get_nested_object
-
-
-def get_wikidata_data(url):
-    response = requests.get(url, headers={'accept': 'application/json',
-                                          'Authorization': f'Bearer {settings.WIKIDATA_API_KEY}'})
-    return response.json() if response.status_code == requests.codes.ok else None
-
-
-def get_wikidata_statements(id):
-    return get_wikidata_data(settings.WIKIDATA_STATEMENTS_URL.format(id))
-
-
-def get_wikidata_label(id, language='en'):
-    return get_wikidata_data(settings.WIKIDATA_LABEL_URL.format(id, language))
+from .wikidata_api import get_wikidata_statements, get_wikidata_label
 
 
 class WikidataSuggestView(AutoResponseView):
